sql 수정

+++ Spring_backend/src/main/java/com/movie/Spring_backend/entity/ReservationEntity.java
@@ -20,12 +20,11 @@ public class ReservationEntity {
     private Date rdate;
 
     @Column(nullable = false)
-    private Long rpice;
+    private Integer rpice;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="miid")
     private MovieInfoEntity miid;
-
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="uid")
@@ -33,7 +32,7 @@ public class ReservationEntity {
 
 
     @Builder
-    public ReservationEntity(Long rid ,Date rdate, Long rpice, MovieInfoEntity miid, MemberEntity uid) {
+    public ReservationEntity(Long rid ,Date rdate, Integer rpice, MovieInfoEntity miid, MemberEntity uid) {
         this.rid=rid;
         this.rdate=rdate;
         this.rpice=rpice;
